Fix Windows build: Skip sigstore signing on Windows
Temporarily skip artifact signing on Windows due to dnspython 2.8.0
dependency issue in sigstore/gh-action-sigstore-python@v3.1.0.

Windows build was failing with:
  ERROR: No matching distribution found for dnspython==2.8.0
  (latest available is 2.7.0)

The Windows binary will still be built and released, just without
sigstore signatures. Linux and macOS binaries will continue to be signed.

This is a temporary workaround until the sigstore action is updated
with correct dependencies.